SDK-691: include update submodules to init&update. resolved cache issue
